+++ src/libs/AxelarGateway.ts
@@ -4,12 +4,13 @@ import {
   CallContractTxArgs,
   CallContractWithTokenTxArgs,
   Environment,
-  EvmChain,
   SendTokenArgs,
 } from "./types";
+
 import erc20Abi from "./abi/erc20Abi.json";
 import { GatewayTx } from "./GatewayTx";
 import { AxelarQueryClient, AxelarQueryClientType } from "./AxelarQueryClient";
+import { EvmChain } from "../constants/EvmChain";
 
 export const AXELAR_GATEWAY: Record<Environment, Partial<Record<EvmChain, string>>> = {
   [Environment.MAINNET]: {
@@ -27,6 +28,8 @@ export const AXELAR_GATEWAY: Record<Environment, Partial<Record<EvmChain, string
     [EvmChain.OPTIMISM]: "0xe432150cce91c13a887f7D836923d5597adD8E31",
     [EvmChain.FILECOIN]: "0xe432150cce91c13a887f7D836923d5597adD8E31",
     [EvmChain.MANTLE]: "0xe432150cce91c13a887f7D836923d5597adD8E31",
+    [EvmChain.SCROLL]: "0xe432150cce91c13a887f7D836923d5597adD8E31",
+    [EvmChain.CENTRIFUGE]: "0xe432150cce91c13a887f7D836923d5597adD8E31",
   },
   [Environment.TESTNET]: {
     [EvmChain.ETHEREUM]: "0xe432150cce91c13a887f7D836923d5597adD8E31",
@@ -45,6 +48,11 @@ export const AXELAR_GATEWAY: Record<Environment, Partial<Record<EvmChain, string
     [EvmChain.LINEA]: "0xe432150cce91c13a887f7D836923d5597adD8E31",
     [EvmChain.FILECOIN]: "0x999117D44220F33e0441fbAb2A5aDB8FF485c54D",
     [EvmChain.MANTLE]: "0xe432150cce91c13a887f7D836923d5597adD8E31",
+    [EvmChain.SCROLL]: "0xe432150cce91c13a887f7D836923d5597adD8E31",
+    [EvmChain.SEPOLIA]: "0xe432150cce91c13a887f7D836923d5597adD8E31",
+    [EvmChain.IMMUTABLE]: "0xe432150cce91c13a887f7D836923d5597adD8E31",
+    [EvmChain.CENTRIFUGE_TESTNET]: "0xe432150cce91c13a887f7D836923d5597adD8E31",
+    [EvmChain.ARBITRUM_SEPOLIA]: "0xe432150cce91c13a887f7D836923d5597adD8E31",
   },
   [Environment.DEVNET]: {
     [EvmChain.ETHEREUM]: "0x7358799e0c8250f0B7D1164824F6Dd5bA31C9Cd6",