+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.13.7] - 2023-DECEMBER-4
+
+- Fix BigNumberUnits utility to adjust for another `fractional component exceeds decimals` edge case
+- Add AddGasToCosmosChain method to AxelarGMPRecoveryAPI
+- Remove singleton pattern that was used in the `getConfigs` function that caused an issue in toggling environments in the AxelarQueryAPI
+- Refactor "EvmChain" enum
+- Update `manualRelayToDestChain` method, adding an additional parameter for `txEventIndex` that is used to confirm events on the network
+- Update `manualRelayToDestChain` method, adding an additional (optional) parameter for `messageId` that is used to recover transactions for GMP transactions from Cosmos source chains
+- Add `Scroll` to testnet/mainnet configs, `Celestia`, `Centrifuge` to testnet/mainnet, `Haqq` to mainnet, `Provenance`, `Sepolia`, `Arbitrum-Sepolia`, `Immutable` to testnet, `Teritori` to mainnet
+- Updated Osmosis testnet to `osmosis-7` that points to `osmo-test-5`
+
 ## [0.13.6] - 2023-SEPTEMBER-1
 
 - Fix the issue with `manualRelayToDestChain` where the web client incorrectly utilizes the provider from the injected browser wallet instead of the rpcUrl embedded in the SDK. This error arises when the injected wallet's network differs from the source chain.